pyproject.toml
[project]
name = "techstax-repo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.1",
    "flask-cors>=6.0.1",
    "flask-pymongo>=3.0.1",
    "gunicorn>=23.0.0",
    "marshmallow>=4.0.0",
    "python-dotenv>=1.1.1",
]
